+++ py/desiutil/install.py
@@ -881,14 +881,14 @@ class DesiInstall(object):
                         # MANIFEST.in and not finding directories.  These
                         # can be ignored.
                         #
-                        manifestre = re.compile(r"no previously-included " +
+                        manifestre = re.compile(r"no (previously-included|) " +
                                                 r"(directories|files) found " +
                                                 r"matching '[^']+'")
                         lines = [l for l in err.split('\n') if len(l) > 0 and
                                  manifestre.search(l) is None]
                         if len(lines) > 0:
                             message = ("Error during installation: " +
-                                       "{0}".format(err))
+                                       "{0}".format("\n".join(lines)))
                             log.critical(message)
                             raise DesiInstallException(message)
             #