+++ doc/changes.rst
@@ -20,7 +20,7 @@ Change Log
 1.3.3 (2016-02-03)
 ------------------
 
-* Added stats module to compute percentiles on distributions.
+* Added :mod:`~desiutil.stats` module to compute percentiles on distributions.
 
 1.3.2 (2016-01-25)
 ------------------
@@ -40,7 +40,7 @@ Change Log
 ------------------
 
 * Updated docstrings for bitmasks.
-* Added :module:`~desuitil.funcfits` module.
+* Added :mod:`~desuitil.funcfits` module.
 
 1.2.0 (2015-11-24)
 ------------------